fixed data type issue

+++ src/dataset_large.py
@@ -238,7 +238,7 @@ class ProtoPlanetaryDisks(Dataset):
       if self.par_norm:
           par = self.MinMaxSc.transform(par.reshape(1, -1))[0]
 
-      return imgs, par
+      return imgs.astype(np.float32), par
 
     def get_dataloader(
         self, batch_size=32, shuffle=True, val_split=0.2, random_seed=42
